Attempt to fix a build failure probably stemming from r2355.
If --disable-network is not given but sdlnet isn't detected, print a warning,
and carry on building as if --disable-network had been given. Don't just
define HAVE_NET before testing if sdlnet is actually installed.

Also apparently the default value for $enable_blah (if you give a --enable-blah
option for an AC_ARG_ENABLE(blah) autoconf directive) is "set" not "yes". But
we don't need to use that at all - we just use ACTION-IF-[NOT-]GIVEN sections.