Make sure all intermission screen time counters reach their targets
Under certain circumstances the intermission screen time counters will
stop before they reach their target[1]. Make sure the counters have the
correct value when the counting finishes so that the intermission screen
has the correct times displayed afterwards.

This fixes the same issue that is covered by prboom-plus r2493.

[1] in particular when compatibility_level < lxdoom_1_compatibility and
both the time and par time are less than the total time. This is for
demo compatibility - you have to stop counting early so that only one
fire/use key press is needed to move on to the next level. (cf. r706)