I_PreInitGraphics: centre a non-fullscreen window
Unconditionally adds "SDL_VIDEO_WINDOW_POS=center" to the environment.
More union { const char *; char *; } abuse so I can pass a const string
to putenv without getting complaints. (putenv's parameter is explicitly
not const; once part of the environment, the string is supposed to be
able to be changed, with changes also becoming part of the environment.)