Build fix: remove the prehistoric USE_SDL macro
SDL has been mandatory (as in, you can't compile the game without it)
for a very long time. Therefore it's pointless to include SDL.h only if
USE_SDL is defined - just include it unconditionally. configure.ac never
bothers to define it - configuration simply fails if SDL is not found.

- r_main.c calls SDL_GetTicks, this used to warn about undeclared
  functions, but otherwise worked.
- v_video.c recently began to use many SDL calls and types, instead of
  abstracting them away into functions within src/SDL, and no longer
  compiles if SDL.h isn't included.
- src/SDL/i_main.c on the other hand doesn't call any SDL functions
  directly, so there's no need to include the header. (Arguably, it
  could be moved out of the SDL subdirectory.)