r_patch.c: FillEmptySpace: work around weapon sprite artifacts
An attempt to work around artifacts on weapon sprites in software mode.

Loop over the patch pixels, looking for edges (transparent pixels next
to solid ones). Copy the solid pixels outwards, dilating the patch's
edges into the surrounding space, creating a halo around it.

Continue until either no transparent pixels remain or some arbitrary
limit is reached. Currently this is eight, although it is likely two
passes would be enough; all we really need is to work around off-by-one
rounding errors in the column drawers after all. Alternatively, the
limit could be omitted, and the whole patch filled in - the code checks
for empty patches that would result in an infinite loop (e.g. STBR127).

Finally, since there are two pixel buffers whose pointers are exchanged
on each pass (as opposed to writing to a scratch buffer and copying back
every time), there must be either an even number of passes, or a final
copy back from the scratch buffer.

Having the space next to the patch's edge match the patch eliminates the
artifacts seen on some weapon sprites at certain software resolutions,
in particular the chainsaw. However it does not fix the line of pixels
at the top of the firing chaingun, which is caused by the column drawer
going past the top of the column and wrapping back to the bottom.